pyproject.toml
[project]
name = "candidate-scanner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ddgs>=9.14.1",
    "duckduckgo-search>=8.1.1",
    "googlesearch-python>=1.3.0",
    "langchain>=1.2.15",
    "langchain-google-genai>=4.2.2",
    "pandas>=3.0.2",
    "python-dotenv>=1.2.2",
    "streamlit>=1.56.0",
]
